Refactored 1.1.3 - contact requests renamed
